Python/Testing with Appium/requirements.txt
Appium-Python-Client==0.30
selenium==3.14.1